refactor(types): KeywordMetadataLoader без unchecked-cast'ов (Sonar)
Закрытие оставшихся 6 issues после первого extract'а:

  S2972 — Builder был внутренним классом ≈110 строк > 25; вынесен в отдельный
    package-private KeywordMetadataBuilder.java со своим набором констант
    FIELD_*/LOCALE_*. Loader стал тонкой обёрткой над ним.

  S1213 — INCLUDE_IN_COMPLETION в GlobalScopeProvider перенесён к
    KEYWORD_CATEGORIES (порядок static-полей по Java code conventions).

  S1309 ×4 — @SuppressWarnings("unchecked") выкошены:
    • root читается типобезопасно через TypeReference<Map<String,Object>>;
    • keywords-узел фильтруется Map.class::isInstance + узкий asEntryMap-cast
      (единственный остающийся SuppressWarnings, локализован комментарием);
    • asStringMap собирает результат проходом по rawMap.entrySet() со
      строгой instanceof-проверкой String-ключа и значения — без cast'ов;
    • readDescriptionByParent теперь строит результат через рекурсивный
      asStringMap, тоже без явных cast'ов.

  Builder ≤ 25 строк на метод, Loader ≤ 95 строк суммарно.

Тесты регистрации и интеграционные — зелёные.
